backend/ingestion_orchestration_fastapi_app/requirements.txt
fastapi
uvicorn[standard]
httpx
qdrant-client~=1.14.0
aiofiles
diskcache
python-multipart
brotli
rawpy
Pillow
exifread
python-xmp-toolkit
umap-learn
psutil

# CUDA-accelerated ML libraries (optional, fallback to CPU)
cuml>=25.02.0; sys_platform != "win32" and platform_machine == "x86_64"
cupy-cuda12x>=12.0.0